+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-Fixed bug whereby sections with empty content were dropped by `markup.iter_markdown_sections`.
+- Fixed bug whereby sections with empty content were dropped by `markup.iter_markdown_sections`.
+- Fixed bug whereby the "booktitle" field was not rendered in the linearization of a `Source` of genre "inbook".
 
 
 ## 3.24.1